Optionally disable animating items when dropping
When this is enabled, items beneath the dropped item re-animate into
place after the user drops something in the list. This feels like a bug,
but I went ahead and made it an option in the interest of not changing
existing behavior for existing users of the library.

Note that the duration is set to 1 because the render animation system
ends up throwing an exception if it's set to 0. I tried replacing the
AnimatedSize widget entirely when this option was set, but swapping an
animated widget in right when we want it to animate means it doesn't
know what its previous size was, so it can't animate. In other words: a
duration of 1 feels wrong, but it's there for a reason.